finish frontend 6/5

+++ owlpot-common/src/main/java/com/owlpot/constant/StatusConstant.java
@@ -6,8 +6,8 @@ package com.owlpot.constant;
 public class StatusConstant {
 
     //启用
-    public static final Integer ENABLE = 1;
+    public static final Byte ENABLE = 1;
 
     //禁用
-    public static final Integer DISABLE = 0;
+    public static final Byte DISABLE = 0;
 }
